HidIO中传入的AimeID全是255时读取aime.txt
