+++ MU3Input/Extensions.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Numerics;
 using System.Runtime.InteropServices;
 using System.Threading;
 
@@ -17,6 +18,19 @@ namespace MU3Input
             {
                 handle.Free();
             }
+        }
+        public static byte[] ToBcd(this BigInteger value)
+        {
+            var length = value.ToString().Length / 2 + value.ToString().Length % 2;
+            byte[] ret = new byte[length];
+            for (int i = length - 1; i >= 0; i--)
+            {
+                ret[i] = (byte)(value % 10);
+                value /= 10;
+                ret[i] |= (byte)((value % 10) << 4);
+                value /= 10;
+            }
+            return ret;
         }
 
     }
