--- MU3Input/Kernel32.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace MU3Input
-{
-    public static class Kernel32
-    {
-
-    }
-}
