+++ MU3Input/Utils.cs
@@ -1,0 +1,82 @@
+﻿using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Numerics;
+using System.Reflection;
+using System.Runtime.InteropServices;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace MU3Input
+{
+    internal class Utils
+    {
+        public static byte[] ReadOrCreateAimeTxt()
+        {
+            byte[] aimeId;
+            var location = Assembly.GetCallingAssembly().Location;
+            string directoryName = Path.GetDirectoryName(location);
+            string deviceDirectory = Path.Combine(directoryName, "DEVICE");
+            string aimeIdPath = Path.Combine(deviceDirectory, "aime.txt");
+            try
+            {
+                var id = BigInteger.Parse(File.ReadAllText(aimeIdPath));
+                var bytes = id.ToBcd();
+                aimeId = new byte[10 - bytes.Length].Concat(bytes).ToArray();
+            }
+            catch (Exception ex)
+            {
+                Random random = new Random();
+                byte[] temp = new byte[10];
+                random.NextBytes(temp);
+                var id = new BigInteger(temp);
+                if (id < -1) id = -(id + 1);
+                id = id % BigInteger.Parse("99999999999999999999");
+                if (!Directory.Exists(deviceDirectory))
+                {
+                    Directory.CreateDirectory(deviceDirectory);
+                }
+                var bytes = id.ToBcd();
+                aimeId = new byte[10 - bytes.Length].Concat(bytes).ToArray();
+                File.WriteAllText(aimeIdPath, id.ToString());
+            }
+            return aimeId;
+        }
+
+        const string defaultIOType = "hid";
+        public static string GetProtocol()
+        {
+            var location = typeof(Mu3IO).Assembly.Location;
+            string directoryName = Path.GetDirectoryName(location);
+            string segatoolsIniPath = Path.Combine(directoryName, "segatools.ini");
+            if (File.Exists(segatoolsIniPath))
+            {
+                StringBuilder temp = new StringBuilder();
+                Utils.GetPrivateProfileString("mu3io", "protocol", defaultIOType, temp, 1024, segatoolsIniPath);
+                return temp.ToString();
+            }
+            return defaultIOType;
+        }
+
+        const int defaultPort = 4354;
+        public static int GetPort()
+        {
+            var location = typeof(Mu3IO).Assembly.Location;
+            string directoryName = Path.GetDirectoryName(location);
+            string segatoolsIniPath = Path.Combine(directoryName, "segatools.ini");
+            if (File.Exists(segatoolsIniPath))
+            {
+                StringBuilder temp = new StringBuilder();
+                Utils.GetPrivateProfileString("mu3io", "port", defaultPort.ToString(), temp, 1024, segatoolsIniPath);
+                if (int.TryParse(temp.ToString(), out int port))
+                {
+                    return port;
+                }
+            }
+            return defaultPort;
+        }
+        [DllImport("kernel32")]//返回取得字符串缓冲区的长度
+        public static extern long GetPrivateProfileString(string section, string key, string def, StringBuilder retVal, int size, string filePath);
+    }
+}
